homeGPT/src/pyproject.toml
[project]
name = "homegpt"
version = "0.1.0"
description = "homeGPT"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "faster-whisper>=1.1.1",
    "openai>=1.97.1",
    "openwakeword>=0.4.0",
    "opuslib>=3.0.1",
    "piper-tts>=1.3.0",
    "pyaudio>=0.2.14",
    "strands-agents>=1.1.0",
    "strands-agents-tools>=0.2.2",
    "webrtcvad>=2.0.10",
]
